resolver: merge name_server_state module into name_server
